Extract to components
